Remove datastore.helpers import in connection.
Done via:
- Moving `_add_keys_to_request` definition to connection module
- Removing only other use (in `batch`, was not a list of keys)
- Copying `_prepare_key_for_request` from helpers (#528 would be
  nice, since we could just remove `_prepare_key_for_request`
  all together)
